TIP-842: Move family translation option model to Akeneo\Pim\Structure
