+++ src/Akeneo/Pim/Structure/Component/Model/FamilyTranslation.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\Entity;
+namespace Akeneo\Pim\Structure\Component\Model;
 
 use Akeneo\Tool\Component\Localization\Model\AbstractTranslation;
 use Akeneo\Pim\Structure\Component\Model\FamilyTranslationInterface;
