+++ src/Pim/Bundle/CatalogBundle/Resources/config/entities.yml
@@ -22,7 +22,7 @@ parameters:
     pim_catalog.entity.attribute_option_value.class:       Pim\Bundle\CatalogBundle\Entity\AttributeOptionValue
     pim_catalog.entity.attribute_requirement.class:        Pim\Bundle\CatalogBundle\Entity\AttributeRequirement
     pim_catalog.entity.family.class:                       Akeneo\Pim\Structure\Component\Model\Family
-    pim_catalog.entity.family_translation.class:           Pim\Bundle\CatalogBundle\Entity\FamilyTranslation
+    pim_catalog.entity.family_translation.class:           Akeneo\Pim\Structure\Component\Model\FamilyTranslation
     pim_catalog.entity.metric.class:                       Pim\Component\Catalog\Model\Metric
     pim_catalog.entity.product_price.class:                Pim\Component\Catalog\Model\ProductPrice
     pim_catalog.entity.family_variant.class:               Pim\Component\Catalog\Model\FamilyVariant
